Factor out integration test commonalities
